Split out threading to its own file
